move providers to testing-lib
