+++ packages/compass-assistant/src/components/assistant-chat.spec.tsx
@@ -28,22 +28,9 @@ import {
   ToolsController,
 } from '@mongodb-js/compass-generative-ai/provider';
 import {
-  AtlasAuthServiceProvider,
-  AtlasServiceProvider,
-  type AtlasAuthService,
-} from '@mongodb-js/atlas-service/provider';
-import {
   ExperimentTestGroups,
   type ExperimentTestGroup,
 } from '@mongodb-js/compass-telemetry';
-
-function WithAtlasProviders({ children }: { children: React.ReactNode }) {
-  return (
-    <AtlasAuthServiceProvider value={{} as AtlasAuthService}>
-      <AtlasServiceProvider>{children}</AtlasServiceProvider>
-    </AtlasAuthServiceProvider>
-  );
-}
 
 describe('AssistantChat', function () {
   const mockMessages: AssistantMessage[] = [
@@ -117,13 +104,11 @@ describe('AssistantChat', function () {
       ensureOptInAndSend: ensureOptInAndSendStub,
     };
     const result = render(
-      <WithAtlasProviders>
-        <ToolsControllerProvider>
-          <AssistantActionsContext.Provider value={assistantActionsContext}>
-            <AssistantChat chat={chat} hasNonGenuineConnections={false} />
-          </AssistantActionsContext.Provider>
-        </ToolsControllerProvider>
-      </WithAtlasProviders>,
+      <ToolsControllerProvider>
+        <AssistantActionsContext.Provider value={assistantActionsContext}>
+          <AssistantChat chat={chat} hasNonGenuineConnections={false} />
+        </AssistantActionsContext.Provider>
+      </ToolsControllerProvider>,
       {
         connections,
         preferences,
@@ -246,11 +231,9 @@ describe('AssistantChat', function () {
     it('shows warning message in chat when connected to non-genuine MongoDB', function () {
       const chat = createMockChat({ messages: [] });
       render(
-        <WithAtlasProviders>
-          <ToolsControllerProvider>
-            <AssistantChat chat={chat} hasNonGenuineConnections={true} />
-          </ToolsControllerProvider>
-        </WithAtlasProviders>
+        <ToolsControllerProvider>
+          <AssistantChat chat={chat} hasNonGenuineConnections={true} />
+        </ToolsControllerProvider>
       );
 
       expect(chat.messages).to.have.length(1);
@@ -265,11 +248,9 @@ describe('AssistantChat', function () {
     it('does not show warning message when all connections are genuine', function () {
       const chat = createMockChat({ messages: [] });
       render(
-        <WithAtlasProviders>
-          <ToolsControllerProvider>
-            <AssistantChat chat={chat} hasNonGenuineConnections={false} />
-          </ToolsControllerProvider>
-        </WithAtlasProviders>,
+        <ToolsControllerProvider>
+          <AssistantChat chat={chat} hasNonGenuineConnections={false} />
+        </ToolsControllerProvider>,
         {
           connections: [],
         }
@@ -284,11 +265,9 @@ describe('AssistantChat', function () {
     it('warning message is removed when all active connections are changed to genuine', async function () {
       const chat = createMockChat({ messages: [] });
       const { rerender } = render(
-        <WithAtlasProviders>
-          <ToolsControllerProvider>
-            <AssistantChat chat={chat} hasNonGenuineConnections={true} />
-          </ToolsControllerProvider>
-        </WithAtlasProviders>,
+        <ToolsControllerProvider>
+          <AssistantChat chat={chat} hasNonGenuineConnections={true} />
+        </ToolsControllerProvider>,
         {}
       );
 
@@ -299,11 +278,9 @@ describe('AssistantChat', function () {
       ).to.exist;
 
       rerender(
-        <WithAtlasProviders>
-          <ToolsControllerProvider>
-            <AssistantChat chat={chat} hasNonGenuineConnections={false} />
-          </ToolsControllerProvider>
-        </WithAtlasProviders>
+        <ToolsControllerProvider>
+          <AssistantChat chat={chat} hasNonGenuineConnections={false} />
+        </ToolsControllerProvider>
       );
 
       await waitFor(() => {
